pyproject.toml
[tool.poetry]
name = "dotobot"
version = "5.1.0"
description = "Discord but worse"
authors = ["Bram Kreulen"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pony = "^0.7.17"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.9"
py-cord = "^2.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
